wip and todos
