+++ Online/OnlineManager.cs
@@ -264,33 +264,6 @@ namespace RainMeadow
             feeds.RemoveAll(f => f.resource == resource);
         }
 
-        [RPCMethod]
-        public static void DeltaReset(RPCEvent rpcEvent, OnlineResource onlineResource, OnlineEntity.EntityId entity)
-        {
-            if (entity != null)
-            {
-                foreach (var feed in OnlineManager.feeds)
-                {
-                    if (feed.player == rpcEvent.from && feed.entity.id == entity && feed.resource == onlineResource)
-                    {
-                        feed.ResetDeltas();
-                        return;
-                    }
-                }
-            }
-            else
-            {
-                foreach (var subscription in OnlineManager.subscriptions)
-                {
-                    if (subscription.player == rpcEvent.from && subscription.resource == onlineResource)
-                    {
-                        subscription.ResetDeltas();
-                        return;
-                    }
-                }
-            }
-        }
-
         // this smells
         public static OnlineResource ResourceFromIdentifier(string rid)
         {
