+++ Online/RPCs.cs
@@ -1,0 +1,44 @@
+﻿namespace RainMeadow
+{
+    public static class RPCs
+    {
+        [RPCMethod]
+        public static void DeltaReset(RPCEvent rpcEvent, OnlineResource onlineResource, OnlineEntity.EntityId entity)
+        {
+            if (entity != null)
+            {
+                foreach (var feed in OnlineManager.feeds)
+                {
+                    if (feed.player == rpcEvent.from && feed.entity.id == entity && feed.resource == onlineResource)
+                    {
+                        feed.ResetDeltas();
+                        return;
+                    }
+                }
+            }
+            else
+            {
+                foreach (var subscription in OnlineManager.subscriptions)
+                {
+                    if (subscription.player == rpcEvent.from && subscription.resource == onlineResource)
+                    {
+                        subscription.ResetDeltas();
+                        return;
+                    }
+                }
+            }
+        }
+
+        [RPCMethod]
+        public static void AddFood(short add)
+        {
+            ((RWCustom.Custom.rainWorld.processManager.currentMainLoop as RainWorldGame)?.Players[0].realizedCreature as Player).AddFood(add);
+        }
+
+        [RPCMethod]
+        public static void AddQuarterFood()
+        {
+            ((RWCustom.Custom.rainWorld.processManager.currentMainLoop as RainWorldGame)?.Players[0].realizedCreature as Player).AddQuarterFood();
+        }
+    }
+}
